Tidy up on main.py, move CSVEditor to own file
