+++ main.py
@@ -1,3 +1,5 @@
+"""Top 10 Pony Video Squeezer 3000 application."""
+import os, shutil, sys
 import tkinter as tk
 from tkinter import ttk, filedialog
 from modules import (  # Import all the neccesary modules lol
@@ -11,13 +13,7 @@ from modules import (  # Import all the neccesary modules lol
     init,
     uploader_diversity,
 )
-import os
-import shutil
-import csv
-import sys
-
-# Main program to be run
-
+from classes.ui import CSVEditor
 
 def browse_file_csv():
     """Handler for the "Browse" button. Opens a file dialog and sets the global
@@ -46,40 +42,40 @@ def run_checks():
         "outputs/temp_outputs/durations_output.csv",  # output 3
     )
 
-    if duplicate_var.get() == False:
+    if check_vars["duplicate"].get() == False:
         shutil.copyfile(
             "outputs/temp_outputs/titles_output.csv",
             "outputs/temp_outputs/processed.csv",
         )
-    if duplicate_var.get():
+    if check_vars["duplicate"].get():
         duplicate.check_duplicates(
             "outputs/shifted_cells.csv",  # input 1
             "outputs/temp_outputs/titles_output.csv",  # input 2
             "outputs/temp_outputs/processed.csv",  # output
         )
 
-    if blacklist_var.get():
+    if check_vars["blacklist"].get():
         blacklist.check_blacklist(
             "outputs/shifted_cells.csv",  # input 1
             "outputs/temp_outputs/processed.csv",  # input 2
             "outputs/temp_outputs/processed.csv",  # output
         )
 
-    if upload_date_var.get():
+    if check_vars["upload_date"].get():
         upload_date.check_dates(
             "outputs/shifted_cells.csv",  # input 1
             "outputs/temp_outputs/processed.csv",  # input 2
             "outputs/temp_outputs/processed.csv",  # output
         )
 
-    if duration_var.get():
+    if check_vars["duration"].get():
         duration_check.check_duration(
             "outputs/shifted_cells.csv",  # input 1
             "outputs/temp_outputs/processed.csv",  # input 2
             "outputs/temp_outputs/processed.csv",  # output
         )
 
-    if fuzzy_var.get():
+    if check_vars["fuzzy"].get():
         fuzzy_check.fuzzy_match(
             "outputs/temp_outputs/processed.csv",  # input
             "outputs/temp_outputs/titles_output.csv",  # output 1
@@ -87,14 +83,14 @@ def run_checks():
             "outputs/temp_outputs/durations_output.csv",  # output 3
         )
 
-    if uploader_occurrence_var.get():
+    if check_vars["uploader_occurrence"].get():
         uploader_occurence.check_uploader_occurrence(
             "outputs/temp_outputs/uploaders_output.csv",  # input
             "outputs/processed.csv",  # input 2
             "outputs/processed.csv",  # output
         )
 
-    if uploader_diversity_var.get():
+    if check_vars["uploader_diversity"].get():
         uploader_diversity.check_uploader_diversity(
             "outputs/temp_outputs/uploaders_output.csv",  # input
             "outputs/processed.csv",  # input 2
@@ -102,7 +98,7 @@ def run_checks():
         )
 
     # Delete output files if present
-    if debug_var == False:
+    if check_vars["debug"] == False:
         temp_output_file_paths = [
             "outputs/temp_outputs/processed_blacklist.csv",
             "outputs/temp_outputs/processed_duplicates.csv",
@@ -121,124 +117,9 @@ def run_checks():
 
 
 def delete_if_present(filepath):
-    """Delete the given file, if it exists on the filesystem."""
+    """Delete the given file if it exists on the filesystem."""
     if os.path.exists(filepath):
         os.remove(filepath)
-
-
-class CSVEditor(tk.Frame):
-    def __init__(self, master):
-        super().__init__(master)
-
-        self.file_path = tk.StringVar()
-        self.data = []
-
-        self.create_widgets()
-
-    def create_widgets(self):
-        # File select button and save button
-        self.browse_button = ttk.Button(self, text="Browse", command=self.browse_file)
-        self.browse_button.pack(pady=5, padx=5, side="top")
-        self.save_button = ttk.Button(
-            self, text="Save Changes", command=self.save_changes
-        )
-        self.save_button.pack(pady=5, padx=5, side="top")
-
-        self.load_button = ttk.Button(
-            self, text="Load processed CSV", command=self.load_processed_csv
-        )
-        self.load_button.pack(pady=5, padx=5, side="top")
-        # Frame to hold the CSV data grid
-        self.data_frame = tk.Frame(self)
-        self.data_frame.pack(expand=True, fill="both", padx=10, pady=10)
-
-        # V-Scrollbar
-        self.scrollbar_y = tk.Scrollbar(self.data_frame, orient="vertical")
-        self.scrollbar_y.pack(side="right", fill="y")
-
-        # H-Scrollbar
-        self.scrollbar_x = tk.Scrollbar(self.data_frame, orient="horizontal")
-        self.scrollbar_x.pack(side="bottom", fill="x")
-
-        # Canvas to hold the data
-        self.canvas = tk.Canvas(
-            self.data_frame,
-            bd=0,
-            xscrollcommand=self.scrollbar_x.set,
-            yscrollcommand=self.scrollbar_y.set,
-            width=1200,
-            height=700,
-        )
-        self.canvas.pack(side="left", fill="both", expand=True)
-        self.scrollbar_x.config(command=self.canvas.xview)
-        self.scrollbar_y.config(command=self.canvas.yview)
-
-        self.inner_frame = tk.Frame(self.canvas)
-        self.canvas.create_window((0, 0), window=self.inner_frame, anchor="nw")
-
-        self.inner_frame.bind("<Configure>", self.on_frame_configure)
-
-    def on_frame_configure(self, event):
-        self.canvas.configure(scrollregion=self.canvas.bbox("all"))
-
-    def browse_file(self):
-        file_path = filedialog.askopenfilename(filetypes=[("CSV Files", "*.csv")])
-        if file_path:
-            self.file_path.set(file_path)
-            self.load_csv(file_path)
-
-    def load_csv(self, file_path):
-        self.data = []
-        with open(file_path, newline="", encoding="utf-8") as csvfile:
-            reader = csv.reader(csvfile)
-            for row in reader:
-                self.data.append(row)
-
-        self.display_csv()
-
-    def load_processed_csv(self):
-        processed_csv_file = "outputs/processed.csv"
-        self.load_csv(processed_csv_file)
-
-    def display_csv(self):
-        # Clear existing widgets
-        for widget in self.inner_frame.winfo_children():
-            widget.destroy()
-
-        # Display CSV data
-        for row_idx, row in enumerate(self.data):
-            for col_idx, value in enumerate(row):
-                entry = tk.Entry(self.inner_frame, width=50)
-                entry.grid(row=row_idx, column=col_idx, padx=5, pady=5)
-                entry.insert(tk.END, value)
-                entry.config(fg="green")
-                if (
-                    "[SIMILARITY DETECTED" in value
-                    or "[DUPLICATE CREATOR]" in value
-                    or "[VIDEO TOO OLD]" in value
-                ):
-                    entry.config(fg="orange")
-                if (
-                    "[5 CHANNEL RULE]" in value
-                    or "[UNSUPPORTED HOST]" in value
-                    or "[VIDEO TOO SHORT]" in value
-                    or "[BLACKLISTED]" in value
-                ):
-                    entry.config(fg="red")
-
-    def save_changes(self):
-        for row_idx, row in enumerate(self.data):
-            for col_idx, _ in enumerate(row):
-                entry_widget = self.inner_frame.grid_slaves(
-                    row=row_idx, column=col_idx
-                )[0]
-                self.data[row_idx][col_idx] = entry_widget.get()
-
-        # Save the changes
-        with open(self.file_path.get(), "w", newline="", encoding="utf-8") as csvfile:
-            writer = csv.writer(csvfile)
-            writer.writerows(self.data)
-
 
 # Create GUI
 root = tk.Tk()
@@ -261,52 +142,33 @@ entry.pack(padx=10, pady=10)
 browse_button = ttk.Button(main_frame, text="Browse", command=browse_file_csv)
 browse_button.pack(pady=10)
 
-# Create checkboxes
-debug_var = tk.BooleanVar()
-debug_checkbox = ttk.Checkbutton(
-    main_frame, text="Enable Debug Files (Broken LOL)", variable=debug_var
-)
-debug_checkbox.pack(pady=40)
+# Create checkboxes and the variables bound to them.
+check_labels = {
+    'debug': 'Enable Debug Files (Broken LOL)',
+    'duplicate': 'Duplicate Check',
+    'blacklist': 'Blacklist Check',
+    'upload_date': 'Upload Date Check',
+    'duration': 'Duration Check',
+    'fuzzy': 'Fuzzy Check',
+    'uploader_occurrence': 'Uploader Occurrence Check',
+    'uploader_diversity': 'Uploader Diversity Check',
+}
 
-duplicate_var = tk.BooleanVar(value=True)
-duplicate_checkbox = ttk.Checkbutton(
-    main_frame, text="Duplicate Check", variable=duplicate_var
-)
-duplicate_checkbox.pack()
+check_vars = {key: tk.BooleanVar(value=True) for key in check_labels}
 
-blacklist_var = tk.BooleanVar(value=True)
-blacklist_checkbox = ttk.Checkbutton(
-    main_frame, text="Blacklist Check", variable=blacklist_var
-)
-blacklist_checkbox.pack()
+# Disable the debug checkbox by default.
+check_vars['debug'] = tk.BooleanVar()
 
-upload_date_var = tk.BooleanVar(value=True)
-upload_date_checkbox = ttk.Checkbutton(
-    main_frame, text="Upload Date Check", variable=upload_date_var
-)
-upload_date_checkbox.pack()
+checkboxes = {
+    key: ttk.Checkbutton(main_frame, text=check_labels[key], variable=check_vars[key])
+    for key in check_labels
+}
 
-duration_var = tk.BooleanVar(value=True)
-duration_checkbox = ttk.Checkbutton(
-    main_frame, text="Duration Check", variable=duration_var
-)
-duration_checkbox.pack()
-
-fuzzy_var = tk.BooleanVar(value=True)
-fuzzy_checkbox = ttk.Checkbutton(main_frame, text="Fuzzy Check", variable=fuzzy_var)
-fuzzy_checkbox.pack()
-
-uploader_occurrence_var = tk.BooleanVar(value=True)
-uploader_occurrence_checkbox = ttk.Checkbutton(
-    main_frame, text="Uploader Occurrence Check", variable=uploader_occurrence_var
-)
-uploader_occurrence_checkbox.pack()
-
-uploader_diversity_var = tk.BooleanVar(value=True)
-uploader_diversity_checkbox = ttk.Checkbutton(
-    main_frame, text="Uploader Diversity Check", variable=uploader_diversity_var
-)
-uploader_diversity_checkbox.pack()
+for key, checkbox in checkboxes.items():
+    if key == 'debug':
+        checkbox.pack(pady=40)
+    else:
+        checkbox.pack()
 
 run_button = ttk.Button(main_frame, text="Run Checks", command=run_checks)
 run_button.pack()
